rename: ai 관련 dto 패키지 이동 TODAK-201
