+++ src/main/java/com/heartsave/todaktodak_api/diary/entity/DiaryEntity.java
@@ -3,7 +3,7 @@ package com.heartsave.todaktodak_api.diary.entity;
 import static com.heartsave.todaktodak_api.diary.constant.DiaryContentConstraintConstant.DIARY_CONTENT_MAX_SIZE;
 import static com.heartsave.todaktodak_api.diary.constant.DiaryContentConstraintConstant.DIARY_CONTENT_MIN_SIZE;
 
-import com.heartsave.todaktodak_api.ai.dto.AiContentResponse;
+import com.heartsave.todaktodak_api.ai.dto.response.AiContentResponse;
 import com.heartsave.todaktodak_api.common.entity.BaseEntity;
 import com.heartsave.todaktodak_api.diary.constant.DiaryEmotion;
 import com.heartsave.todaktodak_api.member.entity.MemberEntity;
